requirements.txt
# Strands Agents SDK
strands-agents>=0.1.0

# AWS Integration
boto3>=1.34.0
botocore>=1.34.0

# HTTP Client for External Services
httpx>=0.27.0
aiohttp>=3.9.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Validation and Models
pydantic>=2.0.0

# Web Search Tools
ddgs>=6.3.0

# Authentication
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# Logging and Observability
structlog>=24.0.0

# Development Tools
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.3.0
mypy>=1.8.0
